test(event): move input value unit tests to value_test.rs
