+++ src/input/event/mod.rs
@@ -4,6 +4,9 @@ pub mod evdev;
 pub mod native;
 pub mod value;
 
+#[cfg(test)]
+pub mod value_test;
+
 /// Events are events that flow from source devices to target devices
 /// TODO: Remove this enum in favor of directly using NativeEvent
 #[derive(Debug, Clone)]
